+++ testtools/tests/__init__.py
@@ -10,20 +10,17 @@ import testscenarios
 def test_suite():
     from testtools.tests import (
         matchers,
+        twistedsupport,
         test_assert_that,
         test_compat,
         test_content,
         test_content_type,
-        test_deferred,
-        test_deferredmatchers,
-        test_deferredruntest,
         test_distutilscmd,
         test_fixturesupport,
         test_helpers,
         test_monkey,
         test_run,
         test_runtest,
-        test_spinner,
         test_tags,
         test_testcase,
         test_testresult,
@@ -32,20 +29,17 @@ def test_suite():
         )
     modules = [
         matchers,
+        twistedsupport,
         test_assert_that,
         test_compat,
         test_content,
         test_content_type,
-        test_deferred,
-        test_deferredmatchers,
-        test_deferredruntest,
         test_distutilscmd,
         test_fixturesupport,
         test_helpers,
         test_monkey,
         test_run,
         test_runtest,
-        test_spinner,
         test_tags,
         test_testcase,
         test_testresult,
