+++ testtools/twistedsupport/__init__.py
@@ -17,13 +17,13 @@ __all__ = [
     'flush_logged_errors',
 ]
 
-from testtools._deferredmatchers import (
+from ._matchers import (
     succeeded,
     failed,
     has_no_result,
 )
 
-from testtools.deferredruntest import (
+from ._runtest import (
     AsynchronousDeferredRunTest,
     AsynchronousDeferredRunTestForBrokenTwisted,
     SynchronousDeferredRunTest,
